requirements.txt
psycopg2-binary==2.9.7
Faker==19.6.2
python-dotenv==1.0.0 